Fixed lack of rounding for parse time

+++ db/migrate/20151027070952_create_parking_meters.rb
@@ -4,8 +4,8 @@ class CreateParkingMeters < ActiveRecord::Migration
       t.integer :name
       t.float :price
       t.float :max_time
-      t.float :start_time
-      t.float :end_time
+      t.integer :start_time
+      t.integer :end_time
       t.boolean :is_broken, :default => false
       t.boolean :is_occupied, :default => false
